requirements.txt
opencv-python==4.9.0.80
mss==9.0.1
numpy==1.26.4
ffmpeg-python==0.2.0
faster-whisper==1.0.1
ultralytics==8.1.24
python-docx==1.1.0
openai==1.30.1
hf_xet==1.1.3
sounddevice==0.5.2
scipy~=1.15.3
SpeechRecognition~=3.14.3
requests~=2.32.4
PyAudio~=0.2.14
keyboard~=0.13.5
streamlit~=1.46.0
python-dotenv~=1.1.0